for #2567, add SelectItemsEngine
