+++ sharding-core/sharding-core-optimize/src/main/java/org/apache/shardingsphere/core/optimize/statement/sharding/dml/select/item/SelectItems.java
@@ -1,0 +1,72 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.shardingsphere.core.optimize.statement.sharding.dml.select.item;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.apache.shardingsphere.core.parse.sql.context.selectitem.SelectItem;
+import org.apache.shardingsphere.core.parse.sql.context.selectitem.StarSelectItem;
+
+import java.util.Collection;
+import java.util.LinkedHashSet;
+import java.util.LinkedList;
+
+/**
+ * Select items.
+ *
+ * @author zhangliang
+ */
+@Getter
+@Setter
+public final class SelectItems {
+    
+    private final Collection<SelectItem> items = new LinkedHashSet<>();
+    
+    private boolean containStar;
+    
+    private int selectListStopIndex;
+    
+    /**
+     * Judge has unqualified star select item.
+     *
+     * @return star select item without owner
+     */
+    public boolean hasUnqualifiedStarSelectItem() {
+        for (SelectItem each : items) {
+            if (each instanceof StarSelectItem && !((StarSelectItem) each).getOwner().isPresent()) {
+                return true;
+            }
+        }
+        return false;
+    }
+    
+    /**
+     * Get qualified star select items.
+     *
+     * @return qualified star select items
+     */
+    public Collection<StarSelectItem> getQualifiedStarSelectItems() {
+        Collection<StarSelectItem> result = new LinkedList<>();
+        for (SelectItem each : items) {
+            if (each instanceof StarSelectItem && ((StarSelectItem) each).getOwner().isPresent()) {
+                result.add((StarSelectItem) each);
+            }
+        }
+        return result;
+    }
+}
